made renovations in hw-23& improved and extracted functions& made js modules before using gulp
