+++ hw-23/index.html
@@ -3,12 +3,13 @@
 <head>
    <meta charset="UTF-8">
    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-   <link rel="stylesheet" href="style.css">
+   <link rel="stylesheet" href="css/reset.css">
+   <link rel="stylesheet" href="css/style.css">
    <link rel="preconnect" href="https://fonts.googleapis.com">
    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
    <link href="https://fonts.googleapis.com/css2?family=Montserrat:wght@300;400;500;600;800&family=Poppins:wght@200;300;500&display=swap" rel="stylesheet">
-   <script src="index.js" type="module" defer></script>
-   <title>SPA shop</title>
+   <script src="js/index.js" type="module" defer></script>
+   <title>Shop</title>
 </head>
 <body>
    <div class="container">
